Rename module 'boot' to 'application'. Remove redundant mvn wrappers from child modules.
