+++ application/src/main/java/ua/com/autokis/application/BootApplication.java
@@ -1,0 +1,14 @@
+package ua.com.autokis.application;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication(scanBasePackages = "ua.com.autokis")
+@EnableFeignClients(basePackages = "ua.com.autokis")
+public class BootApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(BootApplication.class, args);
+    }
+
+}
