+++ application/src/main/java/ua/com/autokis/application/provider/ApiDataProvider.java
@@ -1,0 +1,8 @@
+package ua.com.autokis.application.provider;
+
+import ua.com.autokis.openapi.model.Product;
+import java.util.List;
+
+public interface ApiDataProvider {
+    List<Product> getAllProductsFromAPI();
+}
