+++ application/src/test/java/ua/com/autokis/application/BootApplicationTests.java
@@ -1,0 +1,13 @@
+package ua.com.autokis.application;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BootApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
