--- boot/src/main/java/ua/com/autokis/boot/BootApplication.java
@@ -1,14 +1,0 @@
-package ua.com.autokis.boot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@SpringBootApplication(scanBasePackages = "ua.com.autokis")
-@EnableFeignClients(basePackages = "ua.com.autokis")
-public class BootApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(BootApplication.class, args);
-    }
-
-}
