--- boot/src/main/java/ua/com/autokis/boot/config/ApiTokenConfiguration.java
@@ -1,12 +1,0 @@
-package ua.com.autokis.boot.config;
-
-import lombok.Getter;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@Getter
-public class ApiTokenConfiguration {
-    @Value("${api.token.dd}")
-    private String ddTuningApiToken;
-}
