--- boot/src/main/java/ua/com/autokis/boot/provider/ApiDataProvider.java
@@ -1,8 +1,0 @@
-package ua.com.autokis.boot.provider;
-
-import ua.com.autokis.openapi.model.Product;
-import java.util.List;
-
-public interface ApiDataProvider {
-    List<Product> getAllProductsFromAPI();
-}
