--- boot/src/main/java/ua/com/autokis/boot/service/ProductServiceImpl.java
@@ -1,21 +1,0 @@
-package ua.com.autokis.boot.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Service;
-import ua.com.autokis.boot.provider.ApiDataProvider;
-import ua.com.autokis.openapi.api.ProductsApiDelegate;
-import ua.com.autokis.openapi.model.Product;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class ProductServiceImpl implements ProductsApiDelegate {
-    private final ApiDataProvider dataProvider;
-    @Override
-    public ResponseEntity<List<Product>> getAllProducts() {
-        //Later would use db call
-        return ResponseEntity.ok().body(dataProvider.getAllProductsFromAPI());
-    }
-}
