--- boot/src/test/java/ua/com/autokis/boot/BootApplicationTests.java
@@ -1,13 +1,0 @@
-package ua.com.autokis.boot;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BootApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
